requirements.txt
python-dotenv==1.0.1
streamlit==1.42.0
openai==1.97.0
PyMuPDF==1.26.3
pillow==11.3.0
langchain-openai==0.3.28
langchain-core==0.3.69
langchain==0.3.26
langchain-community==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.5.3

blosc2==3.6.1
Cython==3.1.2
PyQt5==5.15.11
PyQtWebEngine==5.15.7
black==25.1.0